// frontend/src/redux/tag-types.ts
export enum tagTypes {
  admin = "admin",
  user = "user",
  subscriptionPlan = "subscriptionPlan",
  subscription = "subscription",
  contextualize = "contextualize",
  conversation = "conversation",
  analytics = "analytics",
  credit = "credit",
  settings = "settings",
}

export const tagTypesList = [
  tagTypes.admin,
  tagTypes.user,
  tagTypes.subscriptionPlan,
  tagTypes.subscription,
  tagTypes.contextualize,
  tagTypes.conversation,
  tagTypes.analytics,
  tagTypes.credit,
  tagTypes.settings,
];